use absl for int128

+++ clickhouse/columns/date.cpp
@@ -148,7 +148,8 @@ void ColumnDateTime64::Append(const Int64& value) {
 //}
 
 Int64 ColumnDateTime64::At(size_t n) const {
-    return data_->At(n);
+    // make sure to use Absl's Int128 conversion
+    return static_cast<Int64>(data_->At(n));
 }
 
 std::string ColumnDateTime64::Timezone() const {
